Add files via upload
Current Submission is for the Easy portion of the VBA homework. I tried  working on the "moderate" section of the homework and was stuck at the section where the open value was incremented based on the loop criteria. The current logic I have in place is to read the open value of the first record or read the value for date 01, month 01 and store it in open. Then process the condition till ticker value is updated and write the value of close value of 12/31 to close. Calculate Year Change based on Open-Close